# projects/phoenix-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phoenix Platform API
  description: |
    Phoenix Platform API for managing observability cost optimization through
    experiments, pipeline deployments, and metrics analysis.
  version: 2.0.0
  contact:
    name: Phoenix Team
    email: phoenix@example.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://phoenix.example.com/api/v1
    description: Production
security:
  - bearerAuth: []
paths:
...
  /pipelines/deployments:
    get:
      summary: List pipeline deployments
      operationId: listDeployments
      parameters:
        - name: namespace
          in: query
          schema:
            type: string
        - name: pipeline
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployments:
                    type: array
                    items:
                      $ref: '#/components/schemas/PipelineDeployment'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer

    post:
      summary: Create pipeline deployment
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeploymentRequest'
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDeployment'
  /pipelines/deployments/{id}:
    get:
      summary: Get deployment details
      operationId: getDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineDeployment'

    put:
      summary: Update deployment
      operationId: updateDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeploymentRequest'
      responses:
        '204':
          description: Deployment updated

    delete:
      summary: Delete deployment
      operationId: deleteDeployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Deletion accepted
...
  /pipelines/deployments/{id}/versions:
    get:
      summary: List deployment versions
      operationId: listDeploymentVersions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  deployment_id:
                    type: string
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/DeploymentVersion'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PipelineDeployment:
      type: object
      properties:
        id:
          type: string
        deployment_name:
          type: string
        pipeline_name:
          type: string
        namespace:
          type: string
        variant:
          type: string
        target_nodes:
          type: object
          additionalProperties:
            type: string
        parameters:
          type: object
          additionalProperties: true
        resources:
          $ref: '#/components/schemas/ResourceRequirements'
        status:
          type: string
          enum: [pending, deploying, ready, failed, deleting, degraded, healthy]
        phase:
          type: string
        metrics:
          $ref: '#/components/schemas/DeploymentMetrics'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
    CreateDeploymentRequest:
      type: object
      required:
        - deployment_name
        - pipeline_name
        - target_nodes
      properties:
        deployment_name:
          type: string
        pipeline_name:
          type: string
        namespace:
          type: string
          default: default
        variant:
          type: string
        target_nodes:
          type: object
          additionalProperties:
            type: string
        parameters:
          type: object
          additionalProperties: true
        resources:
          $ref: '#/components/schemas/ResourceRequirements'
    UpdateDeploymentRequest:
      type: object
      properties:
        status:
          type: string
        phase:
          type: string
        metrics:
          $ref: '#/components/schemas/DeploymentMetrics'
        status_message:
          type: string
        updated_by:
          type: string
    ResourceRequirements:
      type: object
      properties:
        cpu:
          type: string
          example: "100m"
        memory:
          type: string
          example: "128Mi"
        storage:
          type: string
          example: "1Gi"
    DeploymentMetrics:
      type: object
      properties:
        metrics_per_second:
          type: number
        cardinality_reduction:
          type: number
        error_rate:
          type: number
        latency_p99:
          type: number
    DeploymentVersion:
      type: object
      properties:
        id:
          type: integer
        deployment_id:
          type: string
        version:
          type: integer
        pipeline_config:
          type: string
        parameters:
          type: object
          additionalProperties: true
        deployed_by:
          type: string
        deployed_at:
          type: string
          format: date-time
        notes:
          type: string